+++ xdg.h
@@ -13,4 +13,5 @@ void xdg_data_dirs_destroy(xdg_data_dirs_t dirs);
 const char *xdg_cache_dir(void);
 
 void xdg_find_programs(
-    const char *terminal, struct application_list *applications);
+    const char *terminal, bool include_actions,
+    struct application_list *applications);